Reverting the hardcoded path of compute_saliency.py
The args.data_path is just pointing to the folder where model,
tokens, and other folders are going to be stored. This hard-coded
path is unnecessary, as it is just hard-coding the parameter.

Code is functional after this change start to end and with
visualization.

Just needs as input the folders:

- model
- tokens

inside a folder which is specified by the optional flag when
running the script: --data-path