Modifies termination condition and negative energy change
The current compute_seriation.py assumes two things:

1. energyChange will never be negative.
2. The number of terms is always higher than the default number of iterations.

Assumption 1. is seeing on the original initialization of maxEnergyChange = 0.0
For the simple example provided in the original package for "bessy the cow" on
tokens.txt, this assumption is violated. The added prints showing the currentEnergyChange
and maxEnergyChange shows as iterations increase, the change becomes negative. In this case,
the maxTerm is never selected (if current_energy_change > maxEnergyChange), and the code throws
an error. To fix this, a "-Inf" replaced the original 0.0 for initialization.

Assumption 2: Even fixing Assumption 1., the code will still crash. This second time, this is due
the code not controlling for the number of terms left in the list to be inspected when attempting
a new iteration (the way the code work is, trying over DEFAULTED_NUM_SERIATED_TERMS iterations to
greedly rank the terms and continuously deleting them from the pool of terms. This was confirmed by
printing the python list variable candidateTerms, and seeing the printed iterations to be attempting
to find a term in a then empty list. To remediate this, the loop which uses the number of iterations now
also checks if there is any term left before continuing the loop.

After these two fixes, the code now runs, and the visualization can be generated even for the toy example.